backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
PyYAML==5.4.1
requests==2.26.0
Werkzeug==2.0.1
GitPython==3.1.24
regex==2023.10.3
APScheduler==3.10.4